refactor(Themes): Merge theme functions into a single file
