+++ package.json
@@ -8,12 +8,12 @@
   "main": "./dist/index.js",
   "scripts": {
     "postinstall": "npm run generate",
-    "generate": "npm run generate:selectors && npm run generate:mathjax && npm run generate:examples && npm run generate:themes",
+    "generate": "npm run generate:selectors && npm run generate:mathjax && npm run generate:examples && npm run update:themes",
     "generate:selectors": "ts-node --files src/generate/selectors.ts",
     "generate:mathjax": "ts-node --files src/generate/mathjax.ts",
     "generate:examples": "ts-node --files src/generate/examples.ts",
-    "generate:themes": "ts-node --files src/generate/themes.ts",
-    "create:theme": "ts-node --files src/generate/theme.ts",
+    "update:themes": "ts-node --files src/generate/themes.ts update",
+    "create:theme": "ts-node --files src/generate/themes.ts create",
     "build": "parcel build ./src/index.ts ./src/{common,themes}/**/*.{css,ts} && tsc --emitDeclarationOnly",
     "dev": "parcel ./src/index.html ./src/examples/*.html --open",
     "docs": "parcel build ./src/index.html ./src/examples/*.html -d docs --public-url .",
